+++ src/components/cell/hexagon_2d_cell.rs
@@ -50,15 +50,14 @@ impl Deref for HexagonCell2d {
 impl Cell for HexagonCell2d {
     type Coordinates = IVec3;
 
+    #[inline]
     fn coords(&self) -> &Self::Coordinates {
         &self.coords
     }
 
+    #[inline]
     fn neighbor_coordinates(&self) -> Vec<Self::Coordinates> {
-        NEIGHBOR_COORDINATES
-            .iter()
-            .map(|c| *c + *self.coords())
-            .collect()
+        NEIGHBOR_COORDINATES.map(|c| c + *self.coords()).to_vec()
     }
 }