+++ src/components/cell/neumann_2d_cell.rs
@@ -48,15 +48,14 @@ impl Deref for NeumannCell2d {
 impl Cell for NeumannCell2d {
     type Coordinates = IVec2;
 
+    #[inline]
     fn coords(&self) -> &Self::Coordinates {
         &self.coords
     }
 
+    #[inline]
     fn neighbor_coordinates(&self) -> Vec<Self::Coordinates> {
-        NEIGHBOR_COORDINATES
-            .iter()
-            .map(|c| *c + *self.coords())
-            .collect()
+        NEIGHBOR_COORDINATES.map(|c| c + *self.coords()).to_vec()
     }
 }